chore: add styleguide
